+++ sync-for-commerce/src/sdk/models/operations/getconfigtextsyncflow.ts
@@ -14,6 +14,12 @@ export class GetConfigTextSyncFlowResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * Success
      */
     @SpeakeasyMetadata({ elemType: shared.Localization })