+++ sync-for-commerce/src/sdk/models/operations/getconfiguration.ts
@@ -28,6 +28,12 @@ export class GetConfigurationResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()